proton-mariadb/proton-rds-mariadb-operator/test/requirements.txt
allure-pytest
PyMySQL
paramiko
requests